added routes logic as controllers, code segregation
